// ui/src/lib/api/generated/model/costInput.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Centoire API
 * AI-powered fashion platform API
 * OpenAPI spec version: 1.0.0
 */
import type { CostInputComplexity } from './costInputComplexity';
import type { CostInputMarket } from './costInputMarket';

export interface CostInput {
  category: string;
  /** @minimum 1 */
  quantity: number;
  fabricType: string;
  complexity: CostInputComplexity;
  market: CostInputMarket;
}